pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-hospital_triage"
version = "0.1.0"
description = "Hospital triage and scheduling benchmark for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m hospital_triage.server.app
server = "hospital_triage.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["hospital_triage", "hospital_triage.server"]
package-dir = { "hospital_triage" = ".", "hospital_triage.server" = "server" }
